WeldingProject/requirements.txt
psycopg2-binary==2.9.10
dj-database-url==2.2.0
asgiref==3.11.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
cbor2==5.7.1
celery>=5.3.0
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==5.2.9
django-environ>=0.11.0
django-cors-headers==4.9.0
whitenoise[brotli]==6.6.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Markdown==3.10
msgpack==1.1.2
packaging==25.0
pandas>=2.0.0
openpyxl>=3.1.0
pillow==12.0.0
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
reportlab==4.2.5
requests>=2.28.0
beautifulsoup4>=4.12.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
uvicorn>=0.30.0
rpds-py==0.30.0
service-identity==24.2.0
sqlparse==0.5.4
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
uritemplate==4.2.0
zope.interface==8.1.1
waitress>=3.0.0
pyinstaller>=6.0.0
pyarmor>=8.0.0